+++ string_calculator/implementation/src/main/java/pl/tzch96/string_calculator/Operators.java
@@ -8,6 +8,7 @@ import java.util.Arrays;
 public class Operators {
 
     private static List<String> supportedOperators = Arrays.asList("+", "-", "*", "/");
+    private static List<String> twoArgumentOperators = Arrays.asList("+", "-", "*", "/");
     private static Map<String, Integer> operatorPrecedence = new HashMap<String, Integer>() {
         {
             put("*", 3);
@@ -33,6 +34,10 @@ public class Operators {
         return supportedOperators.contains(str);
     }
 
+    public static boolean isTwoArgumentOperator(String str) {
+        return twoArgumentOperators.contains(str);
+    }
+
     public static boolean isFunction(String str) {
         return str.matches("[A-Za-z]+");
     }
@@ -40,5 +45,38 @@ public class Operators {
     public static Integer getPrecedence(String op) {
         Integer precedence = operatorPrecedence.get(op);
         return precedence;
+    }
+
+    public static Double calculateTwoArg(String operator, Double leftOperand, Double rightOperand) {
+        char op = operator.charAt(0);
+        Double result = 0.0;
+
+        switch(op) {
+            case '+':
+                result = leftOperand + rightOperand;
+                break;
+            case '-':
+                result = leftOperand - rightOperand;
+                break;
+            case '*':
+                result = leftOperand * rightOperand;
+                break;
+            case '/':
+                try {
+                    if (rightOperand == 0) {
+                        throw new ArithmeticException();
+                    }
+                    result = leftOperand / rightOperand;
+                    break;
+                } catch (Exception e) {
+                    System.out.println("Division by 0");
+                    break;
+                }
+            default:
+                System.out.println("Unsupported operator");
+                break;
+        }
+
+        return result;
     }
 }